Поправлена пагинация + корректировки зщапросов

+++ src/main/java/com/model_store/model/dto/ProductDto.java
@@ -1,5 +1,6 @@
 package com.model_store.model.dto;
 
+import com.model_store.model.base.Product;
 import com.model_store.model.constant.Currency;
 import lombok.Builder;
 import lombok.Data;
@@ -9,7 +10,7 @@ import lombok.Data;
 public class ProductDto {
     private Long id;
     private String name;
-    private Long count;
+    private Integer count;
     private Float price;
     private Currency currency;
     private CategoryDto category;
